cpa-offer-aggregator/requirements.txt
playwright==1.42.0
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0